Text extraction & KMeans Fee Monitoring tool
There are 138 distinct fees in the data set, which makes understanding the patterns and sources of these extra shipping fees difficult to track.
We are building a tool which the company is interested in implementing to solve this issue. By creating a Charge Code Index using test extraction, cluster analysis, and domain knowledge, 
we are creating a system of meaningful relationships between individual fees that will allow practinioners to categorize and monitor these charges.